Move account detail forms into AccountDetails component
