+++ client/src/pages/Account/Account.jsx
@@ -3,7 +3,8 @@ import { isLoggedIn } from 'helpers';
 import { Helmet } from 'react-helmet';
 import { useNavigate } from 'react-router-dom';
 import { useCallback, useEffect, useState } from 'react';
-import { createTheme, ThemeProvider, Box, TextField, Button } from '@mui/material';
+import { createTheme, ThemeProvider } from '@mui/material';
+import AccountDetails from './AccountDetails';
 
 const theme = createTheme({
   palette: {
@@ -42,18 +43,6 @@ function Account() {
     });
   }, [getUserData, navigate]);
 
-  const handleAccountUpdate = () => {
-    return;
-  };
-
-  const handlePasswordUpdate = () => {
-    return;
-  };
-
-  const handleAccountDeletion = () => {
-    return;
-  };
-
   return (
     <>
       <Helmet>
@@ -70,66 +59,7 @@ function Account() {
               <p>Loading...</p>
             </div>
           ) : (
-            <div className="md:w-3/4 mx-auto">
-              <Box className="bg-white p-4 rounded-md">
-                <h3 className="text-xl font-bold pb-2">Account Details</h3>
-                <TextField
-                  id="name"
-                  name="name"
-                  label="Name"
-                  defaultValue={userName}
-                  margin="dense"
-                  fullWidth
-                />
-                <TextField
-                  id="email"
-                  name="email"
-                  label="Email"
-                  defaultValue={userEmail}
-                  margin="dense"
-                  fullWidth
-                />
-                <Button onClick={handleAccountUpdate} color="primary" sx={{ mt: 1 }}>
-                  Update Account
-                </Button>
-              </Box>
-              <Box className="bg-white p-4 rounded-md mt-4">
-                <h3 className="text-xl font-bold pb-2">Update Password</h3>
-                <TextField
-                  id="password"
-                  name="password"
-                  label="Current Password"
-                  defaultValue=""
-                  margin="dense"
-                  fullWidth
-                />
-                <TextField
-                  id="newPassword"
-                  name="newPassword"
-                  label="New Password"
-                  defaultValue=""
-                  margin="dense"
-                  fullWidth
-                />
-                <TextField
-                  id="confirmNewPassword"
-                  name="confirmNewPassword"
-                  label="Confirm New Password"
-                  defaultValue=""
-                  margin="dense"
-                  fullWidth
-                />
-                <Button onClick={handlePasswordUpdate} color="primary" sx={{ mt: 1 }}>
-                  Update Password
-                </Button>
-              </Box>
-              <Box className="bg-white p-4 rounded-md mt-4">
-                <h3 className="text-xl font-bold pb-2">Danger Zone</h3>
-                <Button onClick={handleAccountDeletion} color="primary" sx={{ mt: 1 }}>
-                  Delete Account
-                </Button>
-              </Box>
-            </div>
+            <AccountDetails userName={userName} userEmail={userEmail} />
           )}
         </div>
       </ThemeProvider>
