+++ client/src/pages/Account/index.js
@@ -1,0 +1,1 @@
+export { default as AccountDetails } from './AccountDetails';
